CI and Hack: build 802.15.4 SubG BCF image
I'm sure there is a *right* way to do this, but I don't want to maintain
a special command-line for building properly for this board. It should
build with 802.15.4 SubG support for BeagleConnect Freedom by default,
in my opinion. Please let me know if there is a better place to maintain
this.